Add files via upload
this is a project to show the SQL query on software, studies and programmers table to calculate the aggregate function like sum, datetime, max, min, count, .its just a overview off all the major query it include the details about the table. its will be helpful to evaluate the goodwill.